+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - The inspector for the Dialogue Runner now has a drop down for selecting the saliency strategy.
+- `VariableStorageBehaviour` now supports adding a variable change listener for _all_ variables.
 
 ### Changed